Revert "Transfromation param: convert to composite instead of desiconect"
This reverts commit 0148ed73904f6d2bfdfb83ff631b83b5bb9e9f9e.

That commit was supposed to convert transformation nodes to Composite on
disconnecting, but in fact made it impossible to disconnect them at all
when they are linked (as opposed to just being "converted"). Protecting
user from facing raw transformation value might be good, so hopefully it
will be properly implemented later, but as is it brings more severe bug
that what it fixes.

fixes #438